Added SaveFile function (need to write to text instead of rendStr directly)
It should save the file when changes are made although the functionality doesn't work right now since it is writing to tRow[x]->text rather than rendStr. Instead of simply changing it to rendStr, I plan to enable writing to text and updating rendStr afterwords to make sure tabs are accounted for properly. This involves fixing InsertChar which currently writes to rendStr directly instead of text.